refactor: server manager
